requirements.txt
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.5
cryptography==3.4.8
cycler==0.10.0
dnspython==2.1.0
idna==3.2
kiwisolver==1.3.2
loguru==0.5.3
lxml==4.6.3
matplotlib==3.4.3
numpy==1.21.2
pandas==1.3.3
Pillow==8.3.2
pycparser==2.20
pyparsing==2.4.7
python-dateutil==2.8.2
pytz==2021.1
requests==2.26.0
six==1.16.0
tqdm==4.62.2
urllib3==1.26.6
